Kadainomae

+++ app/Http/Controllers/TasksController.php
@@ -8,35 +8,46 @@ use App\Http\Requests;
 
 use App\Task;  
 
-class tasksController extends Controller
+class TasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+   
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
+    
     public function create()
   {
+if (\Auth::check()){
         $task = new Task;
-
+ if (\Auth::user()->id === $task->user_id){
         return view('tasks.create', [
             'task' => $task,
         ]);
+ }
+        else{
+        return redirect ('/');
     }
+}
+else{
+        return redirect ('/');
+    }
+}
 
     /**
      * Store a newly created resource in storage.
@@ -45,20 +56,18 @@ class tasksController extends Controller
      * @return \Illuminate\Http\Response
      */
 public function store(Request $request)
-    {
+     {
         $this->validate($request, [
-            'status' => 'required|max:10',   // add
             'content' => 'required|max:191',
+            'status' => 'required|max:10',
         ]);
 
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
-        $task = new Task;
-        $task->status = $request->status;    // add
-        $task->content = $request->content;
-        $task->save();
-
-
-        return redirect('/');
+        return redirect()->back();
     }
 
     /**
@@ -67,15 +76,24 @@ public function store(Request $request)
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+     public function show($id)
 {
+if (\Auth::check()){
         $task = Task::find($id);
-
+        if (\Auth::user()->id === $task->user_id){
+            
         return view('tasks.show', [
             'task' => $task,
         ]);
     }
-
+    else{
+        return redirect ('/');
+    }
+}
+else{
+        return redirect ('/');
+    }
+}
     /**
      * Show the form for editing the specified resource.
      *
@@ -84,12 +102,22 @@ public function store(Request $request)
      */
     public function edit($id)
    {
+if (\Auth::check()){
         $task = Task::find($id);
-
+if (\Auth::user()->id === $task->user_id){
         return view('tasks.edit', [
             'task' => $task,
         ]);
     }
+     else{
+        return redirect ('/');
+    }
+}
+else{
+        return redirect ('/');
+    }
+}
+    
 
     /**
      * Update the specified resource in storage.
@@ -122,9 +150,12 @@ public function store(Request $request)
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
-        $task = Task::find($id);
-        $task->delete();
+     {
+        $task = \App\Task::find($id);
+
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
